+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
+### Added
+- Add "FriendsBlock" function to all runtimes.
+
 ### Changed
 - Ensure storage write ops return acks in the same order as inputs.
 - Adjust console path for delete all data operation.